requirements.txt
python-dotenv~=1.0.1
pyTelegramBotAPI==4.16.1
requests~=2.31.0
peewee~=3.17.1